bump support no commit

+++ bump.sh
@@ -15,6 +15,7 @@
 set -euo pipefail
 version=${1:-minor}
 prerelease_tag=${2:-}
+commit=${3:-"true"}
 
 echo "Starting ${version} version bump"
 
@@ -26,15 +27,21 @@ if [ ! -z ${prerelease_tag} ]; then
   args="${args} --prerelease=${prerelease_tag}"
 fi
 
+if [ ${commit} = "false" ]; then
+  args="${args} --skip.commit"
+fi
+
 # Generate CHANGELOG and create a commit (see .versionrc.json)
 npx standard-version ${args}
 
-# I am sorry.
-#
-# I've gone diving through the code of `conventional-changelog` to see if there
-# was a way to configure the string and ultimately I decided that a 'sed' was the simpler
-# way to go.
-sed -i.tmp -e 's/BREAKING CHANGES$/BREAKING CHANGES TO EXPERIMENTAL FEATURES/' CHANGELOG.md
-rm CHANGELOG.md.tmp
-git add CHANGELOG.md
-git commit --amend --no-edit
+if [ ${commit} = "true" ]; then
+  # I am sorry.
+  #
+  # I've gone diving through the code of `conventional-changelog` to see if there
+  # was a way to configure the string and ultimately I decided that a 'sed' was the simpler
+  # way to go.
+  sed -i.tmp -e 's/BREAKING CHANGES$/BREAKING CHANGES TO EXPERIMENTAL FEATURES/' CHANGELOG.md
+  rm CHANGELOG.md.tmp
+  git add CHANGELOG.md
+  git commit --amend --no-edit
+fi
